round robin method for execution of RR algorithm and for cleaner look

+++ Job.hpp
@@ -3,14 +3,14 @@
 
 class Job {
   private:
-    char name;                  // name of the job
-    unsigned short admitted;    // time the job was admitted
-    unsigned short length;      // length of time required for the job
+    char name;                // name of the job
+    unsigned int admitted;    // time the job was admitted
+    unsigned int length;      // length of time required for the job
 
   public:
     
     /* constructor */
-    Job(char name, unsigned short admitted, unsigned short length){
+    Job(char name, unsigned int admitted, unsigned int length){
       this->name = name;
       this->admitted = admitted;
       this->length = length;
@@ -19,9 +19,9 @@ class Job {
     /**
      * @brief simulates processing of a job
      * 
-     * @return short 
+     * @return int
      */
-    short processing() {
+    int processing() {
 
       /* decrements length but returns -1 if Job length is 0*/
       if (length == 0) {return -1;}
@@ -34,8 +34,8 @@ class Job {
      * @brief getter for private variables
      */
     char getName() {return name;}
-    unsigned short getAdmitted() {return admitted;}
-    unsigned short getLength() {return length;}
+    unsigned int getAdmitted() {return admitted;}
+    unsigned int getLength() {return length;}
 };  // end Job
 
 #endif
